Resolve duplicate event publishing error
Using the VSCode debug tool I analysed every flow of the subject observable.

Initially I thought the issue lied in KafkaModule since I made it myself
with minimal reference from the internet. But it wasn't. It couldn't.

After spending some 2 hours staring at the screen pressing f10 and f11
I reached the point where I realised that the issue lies where I declare the
AppModule and connect Kafka to the EventBus.

The double registration was taking place in the `.register()` part.
The EventBus does refer to the providers to register the EventHandlers
automatically.
`register()` is a mechanism to add additional handlers, I think.

By removing the EventHandlers array from the `register()` function of the
EventBus of each service, I resolved the issue.